0.7.2 rev 4774

+++ api/libs/api.reminder.php
@@ -24,6 +24,13 @@ class Reminder {
     protected $AllPhones = array();
 
     /**
+     * Contains all frozen users
+     * 
+     * @var array
+     */
+    protected $AllPassive = array();
+
+    /**
      * Placeholder for UbillingSMS object
      *
      * @var object
@@ -54,6 +61,7 @@ class Reminder {
         $this->LoadAllTemplates();
         $this->LoadRemindLogin();
         $this->LoadPhones();
+        $this->LoadPassive();
         $this->sms = new UbillingSMS();
         $this->money = new FundsFlow();
         $this->money->runDataLoders();
@@ -103,6 +111,28 @@ class Reminder {
         $this->AllTemplates = zb_TemplateGetAllUserData();
     }
 
+    protected function LoadPassive() {
+        $query = "SELECT `login` FROM `users` WHERE `Passive`=1";
+        $data = simple_queryall($query);
+        if (!empty($data)) {
+            foreach ($data as $each) {
+                $this->AllPassive[] = $each['login'];
+            }
+        }
+    }
+
+    protected function FilterPassive($login) {
+        if (!empty($this->AllPassive)) {
+            foreach ($this->AllPassive as $each) {
+                if ($each == $login) {
+                    return(true);
+                } else {
+                    return(false);
+                }
+            }
+        }
+    }
+
     /**
      * Make queue for sms send
      * 
@@ -115,28 +145,30 @@ class Reminder {
 
         foreach ($this->AllLogin as $userLoginData) {
             $eachLogin = $userLoginData['login'];
-            if ($this->money->getOnlineLeftCountFast($eachLogin) <= $LiveDays) {
-                if (!file_exists(self::FLAGPREFIX . $eachLogin)) {
-                    $number = $this->AllPhones[$eachLogin]['mobile'];
-                    if (!empty($number)) {
-                        $number = trim($number);
-                        $number = str_replace($this->AltCfg['REMINDER_PREFIX'], '', $number);
-                        $number = vf($number, 3);
-                        $number = $this->AltCfg['REMINDER_PREFIX'] . $number;
-                        $template = $this->AltCfg['REMINDER_TEMPLATE'];
-                        if (!empty($template)) {
-                            $message = zb_TemplateReplace($eachLogin, $template, $this->AllTemplates);
-                            if (!empty($message)) {
-                                $this->sms->sendSMS($number, $message, false);
-                                file_put_contents(self::FLAGPREFIX . $eachLogin, '');
+            if (!$this->FilterPassive($eachLogin)) {
+                if ($this->money->getOnlineLeftCountFast($eachLogin) <= $LiveDays) {
+                    if (!file_exists(self::FLAGPREFIX . $eachLogin)) {
+                        $number = $this->AllPhones[$eachLogin]['mobile'];
+                        if (!empty($number)) {
+                            $number = trim($number);
+                            $number = str_replace($this->AltCfg['REMINDER_PREFIX'], '', $number);
+                            $number = vf($number, 3);
+                            $number = $this->AltCfg['REMINDER_PREFIX'] . $number;
+                            $template = $this->AltCfg['REMINDER_TEMPLATE'];
+                            if (!empty($template)) {
+                                $message = zb_TemplateReplace($eachLogin, $template, $this->AllTemplates);
+                                if (!empty($message)) {
+                                    $this->sms->sendSMS($number, $message, false);
+                                    file_put_contents(self::FLAGPREFIX . $eachLogin, '');
+                                }
                             }
                         }
                     }
-                }
-            } else {
-                if (file_exists(self::FLAGPREFIX . $eachLogin)) {
-                    if (filemtime(self::FLAGPREFIX . $eachLogin) > $CacheTime) {
-                        unlink(self::FLAGPREFIX . $eachLogin);
+                } else {
+                    if (file_exists(self::FLAGPREFIX . $eachLogin)) {
+                        if (filemtime(self::FLAGPREFIX . $eachLogin) > $CacheTime) {
+                            unlink(self::FLAGPREFIX . $eachLogin);
+                        }
                     }
                 }
             }
